Added new rules and tests
