+++ src/index.spec.ts
@@ -11,7 +11,12 @@ describe('stylelint main config', () => {
 		'customSyntax': 'postcss-scss',
 		'extends': ['stylelint-config-standard'],
 		'ignoreFiles': ['**/*.css'],
-		'plugins': ['stylelint-scss', 'stylelint-order', 'stylelint-declaration-block-no-ignored-properties'],
+		'plugins': [
+			'stylelint-scss',
+			'stylelint-order',
+			'stylelint-declaration-block-no-ignored-properties',
+			'stylelint-declaration-strict-value',
+		],
 		'rules': {
 			'annotation-no-unknown': [
 				true,
@@ -35,7 +40,8 @@ describe('stylelint main config', () => {
 					'ignore': ['with-var-inside'],
 				},
 			],
-			'color-no-hex': true,
+			'color-hex-length': 'long',
+			'custom-property-empty-line-before': 'never',
 			'declaration-block-no-duplicate-properties': true,
 			'declaration-block-no-redundant-longhand-properties': null,
 			'declaration-property-unit-allowed-list': {
@@ -81,6 +87,7 @@ describe('stylelint main config', () => {
 				'width': ['px', '%'],
 			},
 			'declaration-property-value-disallowed-list': {
+				'border': ['/none/'],
 				'display': ['/none/'],
 				'text-align': ['/justify/'],
 			},
@@ -89,49 +96,9 @@ describe('stylelint main config', () => {
 			'function-disallowed-list': ['rgb'],
 			'function-url-no-scheme-relative': true,
 			'function-url-scheme-disallowed-list': ['ftp', '/^http/'],
+			'import-notation': 'string',
+			'media-feature-name-disallowed-list': ['max-width', 'min-width', 'width'],
 			'media-feature-name-no-vendor-prefix': true,
-			'media-feature-name-unit-allowed-list': {
-				'/^border-(top|bottom)-(left|right)-radius/': ['px', '%'],
-				'/^border-(top|right|bottom|left)-width/': ['px'],
-				'/^border-(top|right|bottom|left)/': ['px'],
-				'/^margin/': ['px', '%'],
-				'/^padding/': ['px', '%'],
-				'animation': ['ms'],
-				'animation-duration': ['ms'],
-				'background-position': ['px', '%'],
-				'background-size': ['px', '%'],
-				'border': ['px'],
-				'border-radius': ['px', '%'],
-				'border-width': ['px'],
-				'bottom': ['px', '%'],
-				'box-shadow': ['px'],
-				'column-gap': ['px', '%'],
-				'flex-basis': ['px', '%'],
-				'font-size': [],
-				'gap': ['px', '%'],
-				'grid-auto-columns': ['px', '%', 'fr'],
-				'grid-auto-rows': ['px', '%', 'fr'],
-				'grid-column-gap': ['px', '%'],
-				'grid-gap': ['px', '%'],
-				'grid-row-gap': ['px', '%'],
-				'grid-template-columns': ['px', '%', 'fr'],
-				'grid-template-rows': ['px', '%'],
-				'height': ['px', '%'],
-				'left': ['px', '%'],
-				'line-height': [],
-				'max-height': ['px', '%'],
-				'max-width': ['px', '%'],
-				'min-height': ['px', '%'],
-				'min-width': ['px', '%'],
-				'object-position': ['px', '%'],
-				'outline-width': ['px'],
-				'right': ['px', '%'],
-				'text-shadow': ['px'],
-				'top': ['px', '%'],
-				'transition': ['ms'],
-				'transition-duration': ['ms'],
-				'width': ['px', '%'],
-			},
 			'order/order': [
 				{
 					'name': 'charset',
@@ -1775,11 +1742,25 @@ describe('stylelint main config', () => {
 				},
 			],
 			'plugin/declaration-block-no-ignored-properties': true,
+			'rule-empty-line-before': [
+				'always',
+				{
+					'except': ['first-nested'],
+					'ignore': ['after-comment'],
+				},
+			],
 			'rule-selector-property-disallowed-list': {
 				'/^\\.ri-/': ['font-size'],
 				'/ri\\-/': ['font-size'],
 				'i': ['font-size'],
 			},
+			'scale-unlimited/declaration-strict-value': [
+				['/color/', 'background-color', 'font-family', 'font-size', 'size', 'line-height', 'stroke', 'fill'],
+				{
+					'ignoreFunctions': false,
+					'ignoreValues': ['/^rgba/', 'inherit', 'initial', 'none', 'transparent', '0', '1', '/^url/'],
+				},
+			],
 			'scss/at-else-closing-brace-newline-after': 'always-last-in-chain',
 			'scss/at-else-closing-brace-space-after': 'always-intermediate',
 			'scss/at-else-empty-line-before': 'never',
