+++ src/index.ts
@@ -39,8 +39,6 @@ class Configuration implements Config {
 		'stylelint-declaration-strict-value',
 	];
 	public rules = {
-		'annotation-no-unknown': [true, { ignoreAnnotations: [/default\b/] }],
-
 		/* At-rule */
 		'at-rule-no-unknown': [
 			true,
@@ -54,7 +52,7 @@ class Configuration implements Config {
 
 		/* Color */
 		'color-function-notation': ['legacy', { 'ignore': ['with-var-inside'] }],
-		'color-no-hex': true,
+		'color-hex-length': 'long',
 
 		/* Declaration block */
 		'declaration-block-no-duplicate-properties': true,
@@ -72,7 +70,7 @@ class Configuration implements Config {
 
 		/* Media feature */
 		'media-feature-name-no-vendor-prefix': true,
-		'media-feature-name-unit-allowed-list': RULE_PROPERTY_UNIT_ALLOWED_LIST,
+		'media-feature-name-disallowed-list': ['max-width', 'min-width', 'width'],
 
 		/* Rule */
 		'rule-selector-property-disallowed-list': {
@@ -103,6 +101,8 @@ class Configuration implements Config {
 			'display': ['/none/'],
 			/** @see https://github.com/YozhikM/stylelint-a11y/blob/master/src/rules/no-text-align-justify/README.md */
 			'text-align': ['/justify/'],
+
+			'border': ['/none/'],
 		},
 
 		/* Time */
@@ -119,56 +119,144 @@ class Configuration implements Config {
 			},
 		],
 
+		/** Empty lines */
+		'rule-empty-line-before': [
+			'always',
+			{
+				'except': ['first-nested'],
+				'ignore': ['after-comment'],
+			},
+		],
+		'custom-property-empty-line-before': 'never',
+
 		/* Notation */
 		'font-weight-notation': 'numeric',
+		'annotation-no-unknown': [true, { ignoreAnnotations: [/default\b/] }],
+		'import-notation': 'string',
 
-		/* SCSS */
+		/**
+		 * SCSS
+		 * @see https://github.com/stylelint-scss/stylelint-scss/tree/master
+		 * */
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-extend-no-missing-placeholder/README.md */
 		'scss/at-extend-no-missing-placeholder': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-if-no-null/README.md */
 		'scss/at-if-no-null': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-import-partial-extension/README.md */
 		'scss/at-import-partial-extension': 'never',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-rule-no-unknown/README.md */
 		'scss/at-rule-no-unknown': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/comment-no-empty/README.md */
 		'scss/comment-no-empty': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/declaration-nested-properties-no-divided-groups/README.md */
 		'scss/declaration-nested-properties-no-divided-groups': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/declaration-nested-properties/README.md */
 		'scss/declaration-nested-properties': 'never',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/dollar-variable-no-missing-interpolation/README.md */
 		'scss/dollar-variable-no-missing-interpolation': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/dollar-variable-default/README.md */
 		'scss/dollar-variable-default': [true, { ignore: 'local' }],
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/dollar-variable-no-namespaced-assignment/README.md */
 		'scss/dollar-variable-no-namespaced-assignment': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/function-quote-no-quoted-strings-inside/README.md */
 		'scss/function-quote-no-quoted-strings-inside': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/function-unquote-no-unquoted-strings-inside/README.md */
 		'scss/function-unquote-no-unquoted-strings-inside': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/load-no-partial-leading-underscore/README.md */
 		'scss/load-no-partial-leading-underscore': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/no-duplicate-mixins/README.md */
 		'scss/no-duplicate-mixins': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/no-global-function-names/README.md */
 		'scss/no-global-function-names': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/operator-no-newline-after/README.md */
 		'scss/operator-no-newline-after': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/operator-no-newline-before/README.md */
 		'scss/operator-no-newline-before': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/operator-no-unspacedREADME.md */
 		'scss/operator-no-unspaced': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/selector-no-redundant-nesting-selector/README.md */
 		'scss/selector-no-redundant-nesting-selector': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/selector-no-union-class-name/README.md */
 		'scss/selector-no-union-class-name': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/no-duplicate-dollar-variables/README.md */
 		'scss/no-duplicate-dollar-variables': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-else-closing-brace-newline-after/README.md */
 		'scss/at-else-closing-brace-newline-after': 'always-last-in-chain',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-else-closing-brace-space-after/README.md */
 		'scss/at-else-closing-brace-space-after': 'always-intermediate',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-else-empty-line-before/README.md */
 		'scss/at-else-empty-line-before': 'never',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-else-if-parentheses-space-before/README.md */
 		'scss/at-else-if-parentheses-space-before': 'always',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-function-parentheses-space-before/README.md */
 		'scss/at-function-parentheses-space-before': 'never',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-function-pattern/README.md */
 		'scss/at-function-pattern': [
 			'^(-?[a-z][a-z0-9]*)(-[a-z0-9]+)*$',
 			{
 				message: 'Expected function name to be kebab-case',
 			},
 		],
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-if-closing-brace-newline-after/README.md */
 		'scss/at-if-closing-brace-newline-after': 'always-last-in-chain',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-if-closing-brace-space-after/README.md */
 		'scss/at-if-closing-brace-space-after': 'always-intermediate',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-mixin-argumentless-call-parentheses/README.md */
 		'scss/at-mixin-argumentless-call-parentheses': 'never',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-mixin-parentheses-space-before/README.md */
 		'scss/at-mixin-parentheses-space-before': 'never',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-mixin-pattern/README.md */
 		'scss/at-mixin-pattern': [
 			'^(-?[a-z][a-z0-9]*)(-[a-z0-9]+)*$',
 			{
 				message: 'Expected mixin name to be kebab-case',
 			},
 		],
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/at-rule-conditional-no-parentheses/README.md */
 		'scss/at-rule-conditional-no-parentheses': true,
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/dollar-variable-colon-space-after/README.md */
 		'scss/dollar-variable-colon-space-after': 'always-single-line',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/dollar-variable-colon-space-before/README.md */
 		'scss/dollar-variable-colon-space-before': 'never',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/dollar-variable-first-in-block/README.md */
 		'scss/dollar-variable-first-in-block': [true, { ignore: ['comments', 'imports'] }],
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/dollar-variable-empty-line-before/README.md */
 		'scss/dollar-variable-empty-line-before': [
 			'always',
 			{
@@ -176,12 +264,16 @@ class Configuration implements Config {
 				ignore: ['after-comment', 'inside-single-line-block'],
 			},
 		],
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/dollar-variable-pattern/README.md */
 		'scss/dollar-variable-pattern': [
 			'^(-?[a-z][a-z0-9]*)(-[a-z0-9]+)*$',
 			{
 				message: 'Expected variable to be kebab-case',
 			},
 		],
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/double-slash-comment-empty-line-before/README.md */
 		'scss/double-slash-comment-empty-line-before': [
 			'always',
 			{
@@ -189,7 +281,11 @@ class Configuration implements Config {
 				ignore: ['between-comments', 'stylelint-commands'],
 			},
 		],
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/double-slash-comment-whitespace-inside/README.md */
 		'scss/double-slash-comment-whitespace-inside': 'always',
+
+		/** @see https://github.com/stylelint-scss/stylelint-scss/blob/master/src/rules/percent-placeholder-pattern/README.md */
 		'scss/percent-placeholder-pattern': [
 			'^(-?[a-z][a-z0-9]*)(-[a-z0-9]+)*$',
 			{
@@ -201,7 +297,7 @@ class Configuration implements Config {
 		'scale-unlimited/declaration-strict-value': [
 			['/color/', 'background-color', 'font-family', 'font-size', 'size', 'line-height', 'stroke', 'fill'],
 			{
-				'ignoreValues': ['/^rgba/', 'inherit', 'initial', 'none', 'transparent', '/^rem/', '/em$/', '0', '1', '/^url/'],
+				'ignoreValues': ['/^rgba/', 'inherit', 'initial', 'none', 'transparent', '0', '1', '/^url/'],
 				'ignoreFunctions': false,
 			},
 		],
